fix(inspect): use ssh_command pool; promote list_shallow to core (#71)
ls now reuses the codebase-wide SSH ControlMaster (auto/persist=300) instead
of spawning a fresh connection per inspect call. The shallow find lives in
core::remote::ops as remote_list_shallow alongside the recursive variant, so
new commands inherit the pool without copying ssh args.

Also fix off-by-one: 'while i + 3 <= fields.len()' matches the i += 3 step
symmetrically. Previously a missing trailing NUL on a non-GNU find would
silently drop the last triple.
